chore: Workspace 13.10.1 to 15.2.1 upgrade (#23)
* Generate 13.10.6 migration and install deps

* Generate 14.0.0 migration and install deps

* Execute migrations to 14.0.0

* package-lock sync

* Generate 14.2.1 migrations and install deps

* Execute 14.2.1 migrations

* Fix ark-cacher-node serve and build targets

Executors were not set correctly

* Fix linting target for supremegaming-angular

Converted to eslint as well

* Generate and execute migrations for 14.8.6

* Execute trailing migrations from generating new angular app

Was done to get a fresh copy of the e2e app for supremegamin-angular since that config was no longer working

* Migrate last e2e app to eslint

Remove tslint and codelizer as deps

* Remove .eslintrc in favor of .eslintrc.json

* Generate migration for 15.0.0 and install deps

* Execute 15.0.0 migrations

* Manual upgrade cypress to 10.11.0

* Regenerate supremegaming-angular-e2e with updated cypress

Regeneration rolled back cypress to 10.7.0. Left it at that version.

* Generate migrations for 15.2.1 and install deps

* Executed successful migrations

Removed them from migrations. Still need to run failed which have been left in migrations.json

* Fix lint targets for all apps and libs

* Fix lint errors

* Fix typing and syntax issues on supremegaming-angular libs

* Fix typing issues in ticket bot libs

* Fix supreme discord bot docker build not running

Due to mysql not being included in the output package.json

* Remove whitespace in sample .env files

Causes issues with docker if sample used as a template